Create boilerplate for front page of website.
Made three buttons as a header with about, projects, and contact. Will
add functions that would make it so that the buttons immediately scroll
the page down to the appropriate parts of the website.

Currently have a front page that only says "Dong Woo Lee" and a short
title. Will probably change that to some animation or something more
interesting that would make it more worthy of filling up the entire
page.

Also an issue with the index.js file right now with the import, but does
not pose a serious issue and can be fixed later.